+++ packages/openai/src/responses/openai-responses-prepare-tools.ts
@@ -1,6 +1,6 @@
 import {
   LanguageModelV3CallOptions,
-  LanguageModelV3CallWarning,
+  SharedV3Warning,
   UnsupportedFunctionalityError,
 } from '@ai-sdk/provider';
 import { codeInterpreterArgsSchema } from '../tool/code-interpreter';
@@ -33,12 +33,12 @@ export async function prepareResponsesTools({
     | { type: 'code_interpreter' }
     | { type: 'mcp' }
     | { type: 'image_generation' };
-  toolWarnings: LanguageModelV3CallWarning[];
+  toolWarnings: SharedV3Warning[];
 }> {
   // when the tools array is empty, change it to undefined to prevent errors:
   tools = tools?.length ? tools : undefined;
 
-  const toolWarnings: LanguageModelV3CallWarning[] = [];
+  const toolWarnings: SharedV3Warning[] = [];
 
   if (tools == null) {
     return { tools: undefined, toolChoice: undefined, toolWarnings };
@@ -109,6 +109,7 @@ export async function prepareResponsesTools({
                 args.filters != null
                   ? { allowed_domains: args.filters.allowedDomains }
                   : undefined,
+              external_web_access: args.externalWebAccess,
               search_context_size: args.searchContextSize,
               user_location: args.userLocation,
             });
@@ -197,7 +198,10 @@ export async function prepareResponsesTools({
         break;
       }
       default:
-        toolWarnings.push({ type: 'unsupported-tool', tool });
+        toolWarnings.push({
+          type: 'unsupported',
+          feature: `function tool ${tool}`,
+        });
         break;
     }
   }